moved conversions to SN_Math.java
